+++ lib/kon-core/src/commonMain/kotlin/annotations.kt
@@ -1,7 +1,7 @@
 package dev.petuska.kon
 
 /** Marks DSLs with side effects that build and sets entities */
-@DslMarker public annotation class KONSetterDsl
+@DslMarker @Retention(AnnotationRetention.SOURCE) public annotation class KONSetterDsl
 
 /** Marks DSLs with no side effects that build entities */
-@DslMarker public annotation class KONBuilderDsl
+@DslMarker @Retention(AnnotationRetention.SOURCE) public annotation class KONBuilderDsl