+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.1.1
+## Changes
+* New safe array setter `"key".to[1]` that works with single int arrays
+* New object setter `"key" to {}` to allow consistent usage with value keys 
+* `karr` array builder moved out of `KObject` interface
+
 # v1.1.0
 ## Breaking Changes
 * Removed all value setter extensions in favour of `to`. All values should be changes to use that `"myField" to "value"`